% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_discipline_cn.R
\name{add_discipline_cn}
\alias{add_discipline_cn}
\title{add new column for ESI discipline in chinese}
\usage{
add_discipline_cn(df, var)
}
\arguments{
\item{df}{data.frame}

\item{var}{column name for ESI discipline}
}
\value{
data.frame that adds a new column of ESI discipline in chinese
}
\description{
add new column for ESI discipline in chinese
}
\examples{
df \%>\% add_discipline_cn(discipline)
}
