% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_esi_threshold.R
\name{add_esi_threshold}
\alias{add_esi_threshold}
\title{add threshold value for each ESI discipline}
\usage{
add_esi_threshold(df, discipline, date = "last")
}
\arguments{
\item{df}{data.frame}

\item{discipline}{a column belong to the character vector of 22 ESI disciplines}

\item{date}{the date when the threshold line was published, can be \code{date = "last"}, \code{date = "all"} or the specified \code{date = c("20200326")}}
}
\value{
data.frame that adds a new column of threshold line for the specified date
}
\description{
add threshold value for each ESI discipline
}
\examples{
df \%>\% add_esi_threshold(x, date = "last")
df \%>\% add_esi_threshold(x, date = "all")
df \%>\% add_esi_threshold(x, date = c("20200326"))
}
