% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_high_cited.R
\name{add_high_cited}
\alias{add_high_cited}
\title{add high-cited information for ESI discipline}
\usage{
add_high_cited(.data, univ, discipline, source = NULL, scope = "each year")
}
\arguments{
\item{.data}{data.frame}

\item{univ}{a column containing the university name}

\item{discipline}{a column containing the character vector of 22 ESI disciplines}

\item{source}{a data.frame imported by \code{read_highcited()}}

\item{scope}{can be one of \verb{last year}, \verb{all year} and \verb{each year}}
}
\value{
data.frame that adds some new column about high-cited information for the specified discipline.
}
\description{
add high-cited information for ESI discipline
}
\examples{
df \%>\% add_high_cited(univ, discipline, source = hc, scope = "last year")
df \%>\% add_high_cited(univ, discipline, source = hc, scope = "each year")
df \%>\% add_high_cited(univ, discipline, source = hc, scope = "all year")
}
