% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_is_enter_top.R
\name{add_is_enter_top}
\alias{add_is_enter_top}
\title{add is_enter_top_one_percent information for ESI discipline of university}
\usage{
add_is_enter_top(df, univ, discipline, source = NULL, .keep = FALSE)
}
\arguments{
\item{df}{data.frame}

\item{univ}{a column containing the university name}

\item{discipline}{a column containing the character vector of 22 ESI disciplines}

\item{source}{a data.frame imported by \code{read_top_one_percent()}}

\item{.keep}{The default value is \code{FALSE}, only \code{is_enter_top_one_percent} will be added. When \code{.keep = TRUE}, more information will be added.}
}
\value{
data.frame that adds \code{is_enter_top_one_percent} information for the specified discipline of university.
}
\description{
add is_enter_top_one_percent information for ESI discipline of university
}
\examples{
df \%>\% add_is_enter_top(univ, discipline, source = top)
}
