% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_univ_name_cn.R
\name{add_univ_name_cn}
\alias{add_univ_name_cn}
\title{add new column for university name in chinese}
\usage{
add_univ_name_cn(df, var)
}
\arguments{
\item{df}{data.frame}

\item{var}{column name for university name in english}
}
\value{
data.frame that adds a new column of university name in chinese
}
\description{
add new column for university name in chinese
}
\examples{
df \%>\% add_univ_name_cn(univ)
}
