% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{custom_range}
\alias{custom_range}
\title{obtain the range of vector}
\usage{
custom_range(x)
}
\arguments{
\item{x}{vector}
}
\value{
if \code{min(x) == max(x)}, will return \code{range(x)[1]}, else return character \code{range(x)[1] - range(x)[2]}
}
\description{
obtain the range of vector
}
\examples{
custom_range(x)
}
