% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_addr_to_college_name.R
\name{map_addr_to_college_name}
\alias{map_addr_to_college_name}
\title{map addr to college name}
\usage{
map_addr_to_college_name(addr)
}
\arguments{
\item{addr}{a column contains the addr of author}
}
\value{
data.frame
}
\description{
map addr to college name
}
\examples{
df \%>\% mutate(college_name = map_addr_to_college_name(addr))
}
