% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ownership_extract.R
\name{ownership_extract}
\alias{ownership_extract}
\title{extract the author-ownership information of WoS plain txt}
\usage{
ownership_extract(AF, C1, RP)
}
\arguments{
\item{AF}{is a character vector about \verb{Author Full Name}}

\item{C1}{is a character vector about \verb{Author Address}}

\item{RP}{is a character vector about \verb{Reprint Address}}
}
\value{
return a tibble including five columns(.ownership, .names, .addr, .univ, .coll)
}
\description{
Generally, a data.frame which is imported from Web of Science (WoS) plain txt will include a number of two-character field tags, i.e, column names,
such as \code{AU}, \code{AF}, \code{PY} and more. Their specific meaning are available at (https://images.webofknowledge.com/images/help/WOS/hs_wos_fieldtags.html).
The function of \code{ownership_extract()} is to extract and tidy the author-ownership information from some necessary columns.
}
\examples{

res <- df \%>\%
 rowwise() \%>\%
 mutate(
   info = list(ownership_extract(AF, C1, RP))
 ) \%>\%
 ungroup()

This code store the result in list-column. In tidyverse frame,
we can unnest the list-column and then get a tibble with
one row per ownership information * paper combination.

res \%>\%
 select(UT, info) \%>\%
 unnest(cols = c(info)) \%>\%
 select(UT, .ownership, .names, .univ, .coll)



}
