% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_highcited.R
\name{read_highcited}
\alias{read_highcited}
\title{read highcited files}
\usage{
read_highcited(...)
}
\arguments{
\item{...}{file name (named by organization name), now only support \code{xlsx} file format}
}
\value{
data.frame including five columns (univ, discipline, year, n_paper, n_cited)
}
\description{
read highcited files
}
\examples{
hc <- read_highcited("filename.xlsx")
or
here::here("data") \%>\%
  fs::dir_ls(regexp = "\\\\.xlsx$") \%>\%
  read_highcited()

}
