% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_incites.R
\name{read_incites}
\alias{read_incites}
\title{read incites files}
\usage{
read_incites(...)
}
\arguments{
\item{...}{file name (named by organization name), now only support \code{csv} file format}
}
\value{
data.frame including five columns (univ, discipline, year, n_paper, n_cited)
}
\description{
read incites files
}
\examples{
df <- read_incites("filename.csv")
or
here::here("data") \%>\%
  fs::dir_ls(regexp = "\\\\.csv$") \%>\%
  read_incites()


}
