% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_top_one_percent.R
\name{read_top_one_percent}
\alias{read_top_one_percent}
\title{read top_one_percent files}
\usage{
read_top_one_percent(...)
}
\arguments{
\item{...}{file name (named by organization name), now only support \code{xlsx} file format}
}
\value{
data.frame including six columns (univ, discipline, web_of_science_documents, cites, top_papers, is_enter_top_one_percent)
}
\description{
read top_one_percent files
}
\examples{
hc <- read_top_one_percent("filename.xlsx")
or
here::here("data") \%>\%
  fs::dir_ls(regexp = "\\\\.xlsx$") \%>\%
  read_top_one_percent()

}
