% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_wos.R
\name{read_wos}
\alias{read_wos}
\title{read Web of Science files}
\usage{
read_wos(...)
}
\arguments{
\item{...}{file name (the files must be kept in a folder named by the retrieval organization name), now only support \code{txt} file format}
}
\value{
data.frame including some columns (AU, AF, SO, DE, C1, RP, FU, CR, TC, SN, PY, UT)
}
\description{
read Web of Science files
}
\examples{
df <- read_wos("./Sichuan University/filename.txt")
# or
here::here("Sichuan") \%>\%
  fs::dir_ls(regexp = "\\\\.txt$", recurse = TRUE) \%>\%
  read_wos()


}
