% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidyESI-package.R
\docType{package}
\name{tidyESI-package}
\alias{tidyESI}
\alias{tidyESI-package}
\title{tidyESI: A package to tidy ESI data}
\description{
tidy ESI (Essential Science Indicators) data
}
\author{
\strong{Maintainer}: Minjie Wang \email{wangminjiede@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
