% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_high_cited.R
\name{add_high_cited}
\alias{add_high_cited}
\title{add high-cited information for ESI discipline}
\usage{
add_high_cited(
  df,
  discipline,
  univ = "Sichuan Normal University",
  scope = "each year"
)
}
\arguments{
\item{df}{data.frame, but only support dataset about \verb{Sichuan Normal University} now.}

\item{discipline}{a column belong to the character vector of 22 ESI disciplines}

\item{scope}{can be one of \verb{last year}, \verb{all year} and \verb{each year}}
}
\value{
data.frame that adds some new column about high-cited information for the specified discipline.
}
\description{
add high-cited information for ESI discipline
}
\examples{
df \%>\% add_high_cited(x, scope = "last year")
df \%>\% add_high_cited(x, scope = "each year")
df \%>\% add_high_cited(x, scope = "all year")
}
