% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_highcited.R
\name{read_highcited}
\alias{read_highcited}
\title{read highcited file}
\usage{
read_highcited(flnm)
}
\arguments{
\item{flnm}{file name, only only support \code{xlsx} file format}
}
\value{
data.frame including four columns (discipline, year, n_paper, n_cited)
}
\description{
read highcited file
}
\examples{
hc <- read_highcited("filename.xlsx")

}
