% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_incites.R
\name{read_incites}
\alias{read_incites}
\title{read incites file}
\usage{
read_incites(flnm)
}
\arguments{
\item{flnm}{file name, only only support \code{csv} file format}
}
\value{
data.frame including four columns (discipline, year, n_paper, n_cited)
}
\description{
read incites file
}
\examples{
df <- read_incites("filename.csv")

}
